feat: Automatic semver increment selection
If the semver increment is provided explicitly, this has no effect. Otherwise, all of the commits going into the release will have their messages checked. If any start with "break" or "feat", then it'll increment the minor version. If all of the commit messages start with other supported semantic PR types, it'll default to a patch release.

Commit messages that start with anything other than a supported semantic PR type are ignored. If no commits are found that match the `bump_map` rules, a minor version bump is performed by default.

A major version bump is never performed automatically. It must be selected explicitly.

Closes #4